fix: always show agent switcher and provide create-agent path in UI
The AgentSwitcher was hidden when ≤1 agent existed, making it impossible
to select the sole agent — leaving all agent-dependent tabs inaccessible.

- Always render switcher when ≥1 agent exists (even single-agent installs)
- Auto-select the sole agent when only one exists and none is selected
- Add '+ Create Agent' option to the dropdown on all pages
- Show a 'Create Agent' button when zero agents exist
- Extract CreateAgentModal into shared component for reuse
- Auto-select newly created agents in the store
