+++ inc/Core/Admin/Pages/Agent/assets/react/components/AgentListTab.jsx
@@ -12,8 +12,6 @@
 import { useState } from '@wordpress/element';
 import {
 	Button,
-	Modal,
-	TextControl,
 	Notice,
 	Spinner,
 	__experimentalConfirmDialog as ConfirmDialog,
@@ -25,9 +23,10 @@ import { __ } from '@wordpress/i18n';
  */
 import {
 	useManageAgents,
-	useCreateAgent,
 	useDeleteAgent,
 } from '../queries/agents';
+import CreateAgentModal from '@shared/components/CreateAgentModal';
+import { useAgentStore } from '@shared/stores/agentStore';
 
 /**
  * Status badge colors.
@@ -61,96 +60,6 @@ const formatDate = ( dateStr ) => {
 };
 
 /**
- * Create Agent Modal
- */
-const CreateAgentModal = ( { onClose } ) => {
-	const [ slug, setSlug ] = useState( '' );
-	const [ name, setName ] = useState( '' );
-	const [ error, setError ] = useState( '' );
-	const createMutation = useCreateAgent();
-
-	const handleCreate = async () => {
-		setError( '' );
-
-		if ( ! slug.trim() ) {
-			setError( __( 'Agent slug is required.', 'data-machine' ) );
-			return;
-		}
-
-		const result = await createMutation.mutateAsync( {
-			agent_slug: slug.trim(),
-			agent_name: name.trim() || undefined,
-		} );
-
-		if ( result.success ) {
-			onClose();
-		} else {
-			setError(
-				result.message ||
-					__( 'Failed to create agent.', 'data-machine' )
-			);
-		}
-	};
-
-	return (
-		<Modal
-			title={ __( 'Create Agent', 'data-machine' ) }
-			onRequestClose={ onClose }
-			className="datamachine-create-agent-modal"
-		>
-			{ error && (
-				<Notice status="error" isDismissible={ false }>
-					{ error }
-				</Notice>
-			) }
-
-			<TextControl
-				label={ __( 'Slug', 'data-machine' ) }
-				help={ __(
-					'Unique identifier (lowercase, hyphens). Cannot be changed later.',
-					'data-machine'
-				) }
-				value={ slug }
-				onChange={ setSlug }
-				placeholder="my-agent"
-			/>
-
-			<TextControl
-				label={ __( 'Display Name', 'data-machine' ) }
-				help={ __(
-					'Optional. Defaults to slug if empty.',
-					'data-machine'
-				) }
-				value={ name }
-				onChange={ setName }
-				placeholder="My Agent"
-			/>
-
-			<div
-				style={ {
-					display: 'flex',
-					justifyContent: 'flex-end',
-					gap: '8px',
-					marginTop: '16px',
-				} }
-			>
-				<Button variant="tertiary" onClick={ onClose }>
-					{ __( 'Cancel', 'data-machine' ) }
-				</Button>
-				<Button
-					variant="primary"
-					onClick={ handleCreate }
-					isBusy={ createMutation.isPending }
-					disabled={ createMutation.isPending }
-				>
-					{ __( 'Create Agent', 'data-machine' ) }
-				</Button>
-			</div>
-		</Modal>
-	);
-};
-
-/**
  * AgentListTab — main export
  *
  * @param {Object}   props              Component props.
@@ -161,6 +70,7 @@ const AgentListTab = ( { onSelectAgent } ) => {
 	const deleteMutation = useDeleteAgent();
 	const [ showCreateModal, setShowCreateModal ] = useState( false );
 	const [ deleteTarget, setDeleteTarget ] = useState( null );
+	const { setSelectedAgentId } = useAgentStore();
 
 	const handleDelete = async () => {
 		if ( ! deleteTarget ) {
@@ -299,6 +209,11 @@ const AgentListTab = ( { onSelectAgent } ) => {
 			{ showCreateModal && (
 				<CreateAgentModal
 					onClose={ () => setShowCreateModal( false ) }
+					onCreated={ ( agent ) => {
+						if ( agent?.agent_id ) {
+							setSelectedAgentId( agent.agent_id );
+						}
+					} }
 				/>
 			) }
